Add new endpoint - livechat/placeChatOnHold
