+++ ee/app/livechat-enterprise/server/api/index.js
@@ -6,3 +6,4 @@ import './priorities';
 import './tags';
 import './units';
 import './business-hours';
+import './on-hold';
